Add backwards compatibility with http.rb versions < 1.0.0
Since the only API breaking change with older versions of http.rb is the
method call to set headers on the client, and there is value in
supporting older versions (due to caching support, which has been
removed in 1.0.0), it may be reasonable to support both versions.

Tested agains http.rb 0.7.4, 0.9.9 and 1.0.4.