Allow hal_client to be used with http gem version >= 1.0
This patch is based off of previous work by @jeromegn and @tjsousa
(thanks for pointing the way!).

The only change needed to hal_client is to stop using the
Client#with_headers method, and instead make use of Client#headers.

The #headers method was introduced for release 0.8.4 in
https://github.com/httprb/http/commit/a5d22669e5cc59a567e239442a685f2fadca9ed5
and thus it can be safely used even in < 1.0 gem versions.

Earlier PRs: pezra/hal-client#49 jeromegn/hal-client#1